Initial working prototype
There's still a bug where the signal cable renders inside the display;
it seems clipping that might be harder than it's worth for now, so the
current solution was merely to thin the cable out quite a lot. Also, the
case renders bright magenta, which may actually be a missing texture;
I've come to enjoy it, and changed the icon to reflect the color, since
it distinguishes it from a clock well enough.

Displays are devices which instantaneously display signal voltage value,
mapped linearly onto a given minimum/maximum, as an integer with 4
digits of decimal precision. The default of 0/1000 respectively
conveniently maps to the default tachometer setting.