Nearly ready to add clutches
There are some serious bugs with ShaftNetwork.rebuildNetwork at the
moment that preclude this from being release-ready--namely, the
migration to ShaftPart as the unit of storage item (as opposed to
ShaftElement, which represents the whole block) has snuck in some
opportunities to hold references to dead elements (after their holding
block has long been destroyed). This is on the heels of another bug
which caused mass double-counts (or n-counts for n allowed shaft
directions), which show that this code really was not designed for this
kind of consideration. Still, it will have to work, eventually, even as
much as I'm not happy with the current size of rebuildNetwork, nor do I
fully understand the algorithm enough to anticipate its bugs.

Clutches themselves work fine right now, with parameters tuned and
ready. FixedShaft is also added in this commit, and does what it says on
the tin.