breaking interim commit (sorry!)
This introduces StaticShaftNetwork to attempt to rectify the fixed
shaft's implementation and a related clutch bug.

Presently, this tip builds but does not function correctly when
dynamically adding/removing fixed shafts during game time. This will be
fixed, but until then, please use the parent of this commit for testing.