Added Config Copy Tool
Fixes #9.

Works mostly as intended, but there's a conflict in function with the
Wireless Button/Switch that causes it to not trigger for that block in
particular.

Currently supported:
- Lamp Supplies, Wireless Buttons/Switches, and Lamp Sockets copy their
  wireless/power channel names.
- Tachometers, Electrical Probes, Thermal Probes, Data Loggers, and
  Digital Displays copy their units, ranges, and periods, where it makes
  sense.

Other Elements can opt in by implementing the IConfigurable interface as
desired.

Sneak-right-click (or wrench-key-right-click) to copy, and normal
right-click to paste.