Tool handles V sources and beeps
Voltage sources set the key "voltage"; right now, they only validly copy
between each other.

The tool now beeps when used. The beep gives a hint as to whether or not
the change worked: one kind of beep indicates that an upload was
successful, while another indicates that a download was successful; yet
a third indicates failure. The "failure" case is presently invoked on
NodeBase subclasses which don't implement
readConfigTool/writeConfigTool, or for those subclasses that do, when
the contained element does not implement IConfigurable. This is fairly
coarse, and does not necessarily mean that any changes were made--or
that the element actually had any data available to read, for example.