Made joint hubs transmit vertically
This is accomplished in code with a new class which was going to
supplement, not replace, the old joint hubs, but--since these have the
exact same functionality--the ID is reused.

Placing them still requires sneaking, and is a little awkward.

Here's another pony, in case the first doesn't make it:
https://derpicdn.net/img/view/2017/1/16/1340759.png