Add High-Grade Kerosene as a fluid fuel
Higher energy density than kerosene, closer to the max of kerosene-type fuels.
<https://www.engineeringtoolbox.com/fuels-densities-specific-volumes-d_166.html> Highest density from here (840 kg/m^3)
<https://neutrium.net/properties/specific-energy-and-energy-density-of-fuels/> then the 46,2 MJ/kg from here
Gets 38,808 MJ/L
Now, as this is more refined than traditional kerosene, I put it got put at 39,2 because of lower sulfur content reducing polymerization. You can revert to 38,808 if you want, it's slightly closer. It's 99% for my modpack, because I know of no mod that adds it.
<https://polymerdatabase.com/polymer%20chemistry/Vulcanization.html> There's a couple places that mention sulfur extremely helping polymerization. Here's one.