Preliminary fix
This work immediately fixes the most flagrant bug with null SubSystems
in the network when connecting signal cables (and thus closes #194.)

Importantly, this also includes recovery code that allows worlds
corrupted by this bug to be loaded again.

Fixing rendering glitches, possibly also centered on this bug, is not
yet done.