Adds new fluid handler and new thermal converter
I added a new fluid handler that can handle multiple sides of a block
for ease of use where systems may use more than one tank, plus some of
the features from the precise fluid handler that was previously used.

I also added a block that allows IC2 hot coolant to be converted to cold
coolant (and vice versa) as well as water to steam (10x) for steam
turbines from either hot coolant conversion or from other heat sources.

This means that you can now use IC2 fluid reactors to make a lot of
steam power in Electrical Age.

Please note that the thermal converter does explode around 800C, and so
you will want to regulate it with a required signal input to enable.

The granularity of the signal input is poor but will be improved later.