Various fixes to gradle, fixed IProcess issue
I added a few different changes to the readme to reflect the current
status of things a little bit better.

I updated the build.gradle to include more comments, remove unused
dependencies, and fix an issue with Electrical Age server version
checking to be much more strict than before (it never checked). This
should prevent a bunch of server crashes in the future when people use
mismatching versions of different things, but here's to testing that.

I also broke some IProcess consumers by making it Kotlin, since
apparently Kotlin was not smart enough to realize that it can't do the
thing where it overrides the single function, unlike when the code was
written in Java. I blame old versions here...