Increases voltage limits on DC/DC converters
I've limited the secondary on the Grid Transformer to 1/4 of the new top voltage (120kV nominal) but the DC/DC and vDC/DC are not limited in their input voltage due to their flexible nature. I did not touch the T1 poles since T1 Direct Poles are used for more experimental voltage tiers.