A lot of changes to be able to access the signalBusCable.
SideConnectionEnable[] now works with integers, maybe will be reverted.
Is possible to work without that, but will be used to improve the number of connection between the Signal Processor and the cable.

Using the newConnectionAt method to gets the signalBusCable and connect that with then Signal Processor gates.

Only the "A" input is working yet,I don't know how the lrdu and the direction works.