Fixes the version issue if you try to run this from a fork
Semver's developers have a try catch that returns null instead of the exception if it can't parse the crap we provide from RFG; there's no actual "version" information provided, which probably breaks the code. It could also be the `.dirty` is to blame, not root causing that as it's unnecessary.